Replace asm that causes crashing in non VS6 builds
This replaces the inline ASM in non VS6 builds that causes crashing during running of the application.

This is a combination of code suggestions by myself and fixes by xezon.

This fix doesn't include other ASM changes seen in the Zero hour ASM change, this only resolves ASM that is causing crashes.